-Moved the API call, the components now return the info from the events
