+++ client/src/components/ViewEventInfo.js
@@ -1,51 +1,21 @@
-import React, { useEffect, useState } from 'react';
+import React from 'react';
+import { Link} from 'react-router-dom'
 
 // MUI
-import { Button, Grid } from '@material-ui/core';
-import { getEvent } from './../API';
-import { useParams } from 'react-router-dom';
+import { Button, Grid, Typography } from '@material-ui/core';
 
-const input = {
-    title: 'Information about this event:',
-    space: '  ',
-    description:
-        'This is the event description, here you can find a description of the event. In this description, it will be described what the event is about. In this event, you will have access to a series of activities with the purpose of helping you meet new people. Have fun',
-};
+export default function ViewEventInfo({eventDetails}) {
 
-const INITIAL_EVENT_DETAIL = {
-    date: '',
-    isPrivate: false,
-    isVirtual: false,
-    location: '',
-    description: '',
-    image: '',
-    tags: [],
-    organizers: [],
-    attendees: [],
-    posts: [],
-};
-
-export default function ViewEventInfo({}) {
-    const params = useParams();
-    const getEventDetails = (eventId) => {
-        getEvent(eventId)
-            .then(({ status, data }) => {
-                if (status !== 200) {
-                    // 200 indicates successful request
-                    throw new Error('Account not created');
-                }
-                setEventDetails(data);
-                console.log(data);
-            })
-            .catch((error) => console.log(error));
-        return 0;
+    const input = {
+        title: 'What is this event about?',
+        space: '  ',
     };
 
-    useEffect(() => {
-        getEventDetails(params.id);
-    }, []);
+    //<><Typography style={styles.bold}>"I'm bold!" </Typography> </>
+    //const styles = StyleSheet.create({
+    //   bold: {fontWeight: 'bold'}
+    //})
 
-    const [eventDetails, setEventDetails] = useState(INITIAL_EVENT_DETAIL);
     return (
         <>
             <Grid
@@ -53,12 +23,13 @@ export default function ViewEventInfo({}) {
                 direction='row'
                 justify='flex-start'
                 alignItems='center'
+                fontWeight='bold'
                 style={{ margin: 30 }}
                 item
                 xs={10}
             >
-                <Grid item xs={10}>
-                    {eventDetails.description}
+                <Grid style={{padding: '5px 0px'}}>
+                    <Typography variant='h7' component='h1'gutterBottom > {input.title} </Typography>
                 </Grid>
             </Grid>
             <Grid
@@ -84,7 +55,7 @@ export default function ViewEventInfo({}) {
                 xs={10}
             >
                 <Grid item xs={10}>
-                    {input.description}
+                    <Typography variant='h6' component='h4'gutterBottom > {eventDetails.description} </Typography>
                 </Grid>
             </Grid>
             <Grid
@@ -94,10 +65,11 @@ export default function ViewEventInfo({}) {
                 alignItems='flex-end'
             >
                 <Button
+                    component={Link} to="/userHome/"
                     variant='outlined'
                     style={{ backgroundColor: '#6558f5', color: '#ffffff' }}
                 >
-                    Exit
+                    Back
                 </Button>
             </Grid>
         </>
